+++ index.spec.ts
@@ -1,5 +1,4 @@
-import { getConvertedExternals, getEntryPathRegExp } from './index';
-const path = require("path");
+import { getConvertedExternals } from './index';
 
 describe('index', () => {
     describe('getConvertedExternals', () => {
@@ -50,29 +49,6 @@ describe('index', () => {
                     expect(result).toBe(false, `String ${testString} matches some of the regular expressions: ${regExpsExternals.join(', ')}`);
                 });
             });
-        });
-    });
-
-    describe('getEntryPathRegExp', () => {
-        const originalPathJoin = path.join;
-        const entryModule = "index.js";
-
-        afterEach(() => {
-            path.join = originalPathJoin;
-        });
-
-        it('returns RegExp that matches Windows', () => {
-            const appPath = "D:\\Work\\app1\\app";
-            path.join = path.win32.join;
-            const regExp = getEntryPathRegExp(appPath, entryModule);
-            expect(!!regExp.exec(`${appPath}\\${entryModule}`)).toBe(true);
-        });
-
-        it('returns RegExp that works with POSIX paths', () => {
-            const appPath = "/usr/local/lib/app1/app";
-            path.join = path.posix.join;
-            const regExp = getEntryPathRegExp(appPath, entryModule);
-            expect(!!regExp.exec(`${appPath}/${entryModule}`)).toBe(true);
         });
     });
 });